업로드된 파일에 해당하는 gen_file row 생성되도록

+++ src/main/java/com/ll/exam/app10/app/fileUpload/entity/GenFile.java
@@ -15,7 +15,7 @@ import javax.persistence.Entity;
 @ToString(callSuper = true)
 public class GenFile extends BaseEntity {
     private String relTypeCode;
-    private int relId;
+    private long relId;
     private String typeCode;
     private String type2Code;
     private String fileExtTypeCode;
